requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
channels==4.1.0
Django==5.1.1
django-cors-headers==4.4.0
django-nextjs==3.2.0
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
sqlparse==0.5.1
websockets==13.0.1
yarl==1.10.0
